+++ ServerDevcommands/Features/GhostNoSpawns.cs
@@ -10,7 +10,7 @@ namespace ServerDevcommands;
 public class GhostIgnorePlayers {
   private static List<Player> Players() {
     if (!Settings.GhostNoSpawns) return Player.s_players;
-    return Player.s_players.Where(p => !p.InGhostMode()).ToList();
+    return Player.s_players.Where(p => !Ghost.IsGhost(p)).ToList();
   }
 
   static IEnumerable<CodeInstruction> ReplacePlayers(IEnumerable<CodeInstruction> instructions) {
@@ -25,7 +25,7 @@ public class GhostIgnorePlayers {
   [HarmonyPatch(typeof(SpawnSystem), nameof(SpawnSystem.GetPlayersInZone)), HarmonyPostfix]
   static void GhostNoSpawns(List<Player> players) {
     if (!Settings.GhostNoSpawns) return;
-    var toRemove = players.Where(p => p.InGhostMode()).ToList();
+    var toRemove = players.Where(Ghost.IsGhost).ToList();
     foreach (var player in toRemove)
       players.Remove(player);
   }